chore: move mock logout handler to core
