+++ packages/browser/src/ClientAuthentication.spec.ts
@@ -40,7 +40,8 @@ import { fetch } from "@inrupt/universal-fetch";
 import type * as UniversalFetch from "@inrupt/universal-fetch";
 
 import { mockLoginHandler } from "./login/__mocks__/LoginHandler";
-import { mockLogoutHandler } from "./logout/__mocks__/LogoutHandler";
+// eslint-disable-next-line import/no-relative-packages
+import { mockLogoutHandler } from "../../core/src/logout/__mocks__/LogoutHandler";
 import {
   mockSessionInfoManager,
   SessionCreatorCreateResponse,
