+++ packages/node/src/ClientAuthentication.spec.ts
@@ -34,7 +34,8 @@ import type * as UniversalFetch from "@inrupt/universal-fetch";
 import { Session } from "./Session";
 
 import { mockLoginHandler } from "./login/__mocks__/LoginHandler";
-import { mockLogoutHandler } from "./logout/__mocks__/LogoutHandler";
+// eslint-disable-next-line import/no-relative-packages
+import { mockLogoutHandler } from "../../core/src/logout/__mocks__/LogoutHandler";
 import {
   mockSessionInfoManager,
   SessionCreatorCreateResponse,
