+++ packages/node/src/sessionInfo/SessionInfoManager.spec.ts
@@ -22,7 +22,8 @@
 import { jest, it, describe, expect } from "@jest/globals";
 import { mockStorageUtility } from "@inrupt/solid-client-authn-core";
 import { UuidGeneratorMock } from "../util/__mocks__/UuidGenerator";
-import { mockLogoutHandler } from "../logout/__mocks__/LogoutHandler";
+// eslint-disable-next-line import/no-relative-packages
+import { mockLogoutHandler } from "../../../core/src/logout/__mocks__/LogoutHandler";
 import { SessionInfoManager } from "./SessionInfoManager";
 import { KEY_REGISTERED_SESSIONS } from "../constant";
